move picture

+++ src/Types.hs
@@ -2,7 +2,7 @@ module Types where
 
 
 type World = (Pos, State)
-type Pos = (Int, Int)--position x,y
+type Pos = (Float, Float)--position x,y
 type State = Int --0 - empty, 1-black, 2-white
 
 offsetX :: Float
